frontend/chat: scaffolding llm feedback; refactoring
